+++ cmd/hook/hook.go
@@ -17,6 +17,7 @@ func NewHookCommand(client client.API, w io.Writer) *cobra.Command {
 	cmd.AddCommand(
 		NewListCommand("list", client, w),
 		NewEventTypeCommand("event-types", client, w),
+		NewListSecretCommand("secrets", client, w),
 	)
 	return cmd
 }