+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## next
 
+## 0.0.13
+
 - Ability to list your hook secrets via `hook secrets`.
 - Add the ID and WorkspaceID to the secret output.
 - Ability to create a workspace via `workspace create`.